requirements.txt
amqp==2.5.2
attrs==19.3.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.1
billiard==3.6.1.0
cached-property==1.5.1
celery==4.3.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
decorator==4.4.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.6
django-allauth==0.40.0
django-environ==0.4.5
django-extensions==2.2.5
django-filter==2.2.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.10.3
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.17.0
gunicorn==20.0.0
idna==2.8
importlib-metadata==0.23
inflection==0.3.1
ipdb==0.12.2
ipython==7.8.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.2.0
kombu==4.6.6
lxml==4.4.2
Markdown==3.1.1
MarkupSafe==1.1.1
more-itertools==7.2.0
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==19.2
paramiko==2.6.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
psycopg2-binary==2.8.3
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.4.2
PyNaCl==1.3.0
pyparsing==2.4.5
pyrsistent==0.15.6
python3-openid==3.1.0
pytz==2019.2
PyYAML==3.13
redis==3.3.11
requests==2.22.0
requests-cache==0.5.2
requests-oauthlib==1.2.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
Serializer==0.2.1
simplejson==3.16.0
six==1.12.0
soupsieve==1.9.5
sqlparse==0.3.0
texttable==1.6.2
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.25.6
vine==1.3.0
wcwidth==0.1.7
websocket-client==0.56.0
whitenoise==4.1.4
zipp==0.6.0
